Add new test for creating deterministic privateref
