+++ wnfs/src/private/mod.rs
@@ -6,6 +6,7 @@ mod key;
 pub mod namefilter;
 pub mod node;
 mod previous;
+mod privateref;
 
 pub use directory::*;
 pub use file::*;
@@ -15,3 +16,4 @@ pub use key::*;
 pub use namefilter::*;
 pub use node::*;
 pub use previous::*;
+pub use privateref::*;
